Fix update category tab to bottom bar
