+++ app/src/main/java/com/fitmate/fitmate/presentation/ui/login/NicknameFragment.kt
@@ -1,9 +1,7 @@
 package com.fitmate.fitmate.presentation.ui.login
 
-import android.content.Context
 import android.os.Bundle
 import android.view.View
-import android.view.inputmethod.InputMethodManager
 import android.widget.Toast
 import androidx.fragment.app.Fragment
 import androidx.navigation.fragment.findNavController
