+++ app/src/main/res/values/strings.xml
@@ -126,6 +126,7 @@
 
     <string name="menu_home">Home</string>
     <string name="menu_myfit">MyFit</string>
+    <string name="menu_category">Category</string>
     <string name="menu_chat">Chat</string>
     <string name="menu_my_page">My Page</string>
